moved GameAttendance DTos and mappers to separate folder
