+++ src/CoachConnect.BusinessLayer/DTOs/GameAttendances/GameAttendanceRegistrationDTO.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CoachConnect.BusinessLayer.DTOs.Games;
+namespace CoachConnect.BusinessLayer.DTOs.GameAttendances;
 public record GameAttendanceRegistrationDTO(
 
     GameId GameId,
