+++ src/CoachConnect.BusinessLayer/Services/GameAttendanceService.cs
@@ -1,4 +1,4 @@
-﻿using CoachConnect.BusinessLayer.DTOs.Games;
+﻿using CoachConnect.BusinessLayer.DTOs.GameAttendances;
 using CoachConnect.BusinessLayer.Mappers;
 using CoachConnect.BusinessLayer.Mappers.GameMappers;
 using CoachConnect.BusinessLayer.Mappers.Interfaces;
